Split angle output persistence
